+++ quickbi-public-20220101/alibabacloud_quickbi_public20220101/models/_clear_dynamic_tag_cache_request.py
@@ -9,6 +9,8 @@ class ClearDynamicTagCacheRequest(DaraModel):
         self,
         config_id: str = None,
     ):
+        # The ID of the tag configuration.
+        # 
         # This parameter is required.
         self.config_id = config_id