+++ quickbi-public-20220101/alibabacloud_quickbi_public20220101/models/_create_cube_by_sql_request.py
@@ -22,13 +22,13 @@ class CreateCubeBySqlRequest(DaraModel):
         # 
         # This parameter is required.
         self.custom_sql = custom_sql
-        # The data source ID.
+        # The ID of the data source.
         # 
         # This parameter is required.
         self.ds_id = ds_id
         # The placeholder parameters. For more information, see the supplementary description below.
         self.placeholders = placeholders
-        # The Quick BI UserId of a user who has permissions to create datasets. This is not your Alibaba Cloud account ID. You can call the [QueryUserInfoByAccount](https://next.api.aliyun.com/api/quickbi-public/2022-01-01/QueryUserInfoByAccount?spm=api-workbench.api_explorer.0.0.672f50daGq9ooV&params=%7B%7D&tab=DOC&sdkStyle=old&RegionId=cn-hangzhou) operation to obtain the UserId.
+        # The UserId of a Quick BI user who has permissions to create datasets. This is not your Alibaba Cloud account ID. You can call the [QueryUserInfoByAccount](https://next.api.aliyun.com/api/quickbi-public/2022-01-01/QueryUserInfoByAccount?spm=api-workbench.api_explorer.0.0.672f50daGq9ooV&params=%7B%7D&tab=DOC&sdkStyle=old&RegionId=ap-southeast-1) operation to obtain the UserId.
         # 
         # This parameter is required.
         self.user_id = user_id