+++ quickbi-public-20220101/alibabacloud_quickbi_public20220101/models/_query_user_by_mobile_account_request.py
@@ -14,7 +14,7 @@ class QueryUserByMobileAccountRequest(DaraModel):
         # 
         # - DingTalk: ding
         # - WeCom: corp_weixin
-        # - Lark: feishu.
+        # - Lark: feishu
         # 
         # This parameter is required.
         self.mobile_type = mobile_type
@@ -23,7 +23,7 @@ class QueryUserByMobileAccountRequest(DaraModel):
         # - DingTalk: The unionId of the DingTalk account.
         # - WeCom: The userId of the WeCom account.
         # - Lark: The userId of the Lark account.
-        # >Notice: The mobileUserId must be obtained by calling the relevant DingTalk, WeCom, or Lark API operations..
+        # >Notice: The mobileUserId must be obtained by calling the relevant API operations of DingTalk, WeCom, or Lark.
         # 
         # This parameter is required.
         self.mobile_user_id = mobile_user_id