+++ quickbi-public-20220101/alibabacloud_quickbi_public20220101/models/_query_user_list_request.py
@@ -11,18 +11,16 @@ class QueryUserListRequest(DaraModel):
         page_num: int = None,
         page_size: int = None,
     ):
-        # The keyword to search for organization members by username or nickname.
+        # The keyword for the username or nickname of the organization member.
         self.keyword = keyword
-        # The page number to return.
+        # The page number of the organization member list.
         # 
-        # - Starting value: 1
-        # 
+        # - Minimum value: 1
         # - Default value: 1
         self.page_num = page_num
-        # The number of organization members to return per page.
+        # The number of rows per page in a paged query.
         # 
         # - Default value: 10
-        # 
         # - Maximum value: 1000
         self.page_size = page_size