+++ quickbi-public-20220101/alibabacloud_quickbi_public20220101/models/_query_user_list_response_body.py
@@ -16,13 +16,12 @@ class QueryUserListResponseBody(DaraModel):
     ):
         # The request ID.
         self.request_id = request_id
-        # The paginated list of users. The `Data` parameter contains the details of each organization member.
+        # The paginated result of the user list. The detailed information of organization members is stored in the Data response parameter.
         self.result = result
         # Indicates whether the request was successful. Valid values:
         # 
-        # - `true`: The request was successful.
-        # 
-        # - `false`: The request failed.
+        # - true: The request was successful.
+        # - false: The request failed.
         self.success = success
 
     def validate(self):
@@ -68,13 +67,13 @@ class QueryUserListResponseBodyResult(DaraModel):
         total_num: int = None,
         total_pages: int = None,
     ):
-        # The list of users.
+        # The list of users returned by the request.
         self.data = data
         # The page number.
         self.page_num = page_num
-        # The number of entries per page.
+        # The number of rows per page specified in the request.
         self.page_size = page_size
-        # The total number of matching users.
+        # The total number of rows.
         self.total_num = total_num
         # The total number of pages.
         self.total_pages = total_pages
@@ -147,54 +146,51 @@ class QueryUserListResponseBodyResultData(DaraModel):
         user_id: str = None,
         user_type: int = None,
     ):
-        # The Alibaba Cloud account ID. For users not added through RAM, this ID is available only after they log in.
+        # The Alibaba Cloud account ID. For users who are not added through RAM self-service, the Alibaba Cloud ID can only be obtained after they log on.
         self.account_id = account_id
         # The Alibaba Cloud account name.
         self.account_name = account_name
-        # Indicates whether the user is an organization administrator. Valid values:
+        # Indicates whether the user is bound to the organization administrator role. Valid values:
         # 
-        # - `true`: Yes
+        # - true: Yes.
+        # - false: No.
         # 
-        # - `false`: No
-        # 
-        # >Notice: 
-        # 
-        # This parameter is deprecated. Use the `RoleIdList` parameter instead.
+        # <notice>This parameter is deprecated. Use the RoleIdList parameter instead.</notice>
         self.admin_user = admin_user
-        # Indicates whether the user is a permission administrator. Valid values:
+        # Indicates whether the user is attached to the permission management administrator role. Valid values:
         # 
-        # - `true`: Yes
+        # - true: Yes.
+        # - false: No.
         # 
-        # - `false`: No
-        # 
-        # >Notice: 
-        # 
-        # This parameter is deprecated. Use the `RoleIdList` parameter instead.
+        # <notice>This parameter has expired and is no longer recommended. Use the RoleIdList parameter instead.</notice>
         self.auth_admin_user = auth_admin_user
+        # The intelligent module quota of the user.
+        # 
+        # - qreport: Q report
+        # - qExploreNum: Q exploration edition
+        # - smartQAskNum: Q data inquiry
+        # - smartQDevNum: Q builder
         self.copilot_modules = copilot_modules
-        # Indicates whether the user is inactive.
+        # The user status. Valid values: 
         # 
-        # - `false`: Active
-        # 
-        # - `true`: Inactive
+        # - false: active
+        # - true: inactive
         self.is_deleted = is_deleted
-        # The Unix timestamp (in milliseconds) that indicates when the user joined the organization.
+        # The date when the member joined the organization.
         self.joined_date = joined_date
-        # The Unix timestamp (in milliseconds) of the user\\"s last login.
+        # The last logon time.
         self.last_login_time = last_login_time
-        # The nickname of the user.
+        # The nickname of the organization member.
         self.nick_name = nick_name
-        # The IDs of the organization roles assigned to the user.
+        # The list of organization role IDs bound to the user.
         self.role_id_list = role_id_list
         # The user ID in Quick BI.
         self.user_id = user_id
         # The user type of the organization member. Valid values:
         # 
-        # - `1`: developer
-        # 
-        # - `2`: viewer
-        # 
-        # - `3`: analyst
+        # - 1: developer
+        # - 2: visitor
+        # - 3: analyst
         self.user_type = user_type
 
     def validate(self):