+++ quickbi-public-20220101/alibabacloud_quickbi_public20220101/models/_update_cube_by_sql_request.py
@@ -28,7 +28,7 @@ class UpdateCubeBySqlRequest(DaraModel):
         self.ds_id = ds_id
         # The placeholder parameters. For more information, see the supplementary description below.
         self.placeholders = placeholders
-        # The Quick BI UserId of a user who has permissions to create datasets. This is not your Alibaba Cloud account ID. Call the [QueryUserInfoByAccount](https://next.api.aliyun.com/api/quickbi-public/2022-01-01/QueryUserInfoByAccount?spm=api-workbench.api_explorer.0.0.672f50daGq9ooV&params=%7B%7D&tab=DOC&sdkStyle=old&RegionId=cn-hangzhou) operation to obtain the UserId.
+        # The user ID of a Quick BI user who has permissions to create datasets. This is not your Alibaba Cloud account ID. You can call the [QueryUserInfoByAccount](https://next.api.aliyun.com/api/quickbi-public/2022-01-01/QueryUserInfoByAccount?spm=api-workbench.api_explorer.0.0.672f50daGq9ooV&params=%7B%7D&tab=DOC&sdkStyle=old&RegionId=cn-hangzhou) operation to obtain the user ID.
         # 
         # This parameter is required.
         self.user_id = user_id